+++ WPFApp_LibraryManager/Services/CategoryService.cs
@@ -118,20 +118,21 @@ namespace WPFApp_LibraryManager.Services
         {
             try
             {
-                _categoryRepository.DeleteCategory(categoryId);
-
-                MessageBox.Show("Category deleted successfully!");
-            }
-            catch (SqlException ex)
-            {
-                if (ex.Errors[0].Number == 547)
+                if (_categoryRepository.IsCategoryInUse(categoryId) == true)
                 {
-                    MessageBox.Show($"Cannot delete category #{categoryId}. Some books are still associated with it.");
+                    MessageBox.Show($"Could not delete Category #{categoryId}. Some books are still associated with it.");
+                }
+                else
+                {
+
+                    _categoryRepository.DeleteCategory(categoryId);
+
+                    MessageBox.Show($"Category #{categoryId} deleted successfully!");
                 }
             }
             catch (Exception ex)
             {
-                MessageBox.Show($"Could not delete category: {ex.Message}");
+                MessageBox.Show($"Could not delete Category #{categoryId}: {ex.Message}");
             }
         }
     }
