+++ WPFApp_LibraryManager/Services/PublisherService.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
 using System.Windows;
 using WPFApp_LibraryManager.Interfaces;
 using WPFApp_LibraryManager.Models;
@@ -118,20 +117,21 @@ namespace WPFApp_LibraryManager.Services
         {
             try
             {
-                _publisherRepository.DeletePublisher(publisherId);
-
-                MessageBox.Show("Publisher deleted successfully!");
-            }
-            catch (SqlException ex)
-            {
-                if (ex.Errors[0].Number == 547)
+                if (_publisherRepository.IsPublisherInUse(publisherId) == true)
                 {
-                    MessageBox.Show($"Cannot delete publisher #{publisherId}. Some books are still associated with it.");
+                    MessageBox.Show($"Could not delete Publisher #{publisherId}. Some books are still associated with it.");
+                }
+                else
+                {
+
+                    _publisherRepository.DeletePublisher(publisherId);
+
+                    MessageBox.Show($"Publisher #{publisherId} deleted successfully!");
                 }
             }
             catch (Exception ex)
             {
-                MessageBox.Show($"Could not delete publisher: {ex.Message}");
+                MessageBox.Show($"Could not delete Publisher #{publisherId}: {ex.Message}");
             }
         }
     }
